Big changes. 1)  Ripped out the optimizing loop 2)  Added in to the scoring bits designed to optimize by scoring not by     constraining 3)  Now uses the topology description to determine the available     physical network.
It doesn't produce feasible solutions as consistently as the old mechanism
did, but I suspect this just means I need to be a bit smarter with the
scoring, and also look for a global feasible solution and record it
instead of just the best score.  But perhaps the score tweaks will get it.