Loss now has poisson and random table (of rates) Bandwidth now has const rate, uniform, determ and random tables, poisson.
so that's all of the functionality that people wanted.
---

I don't think poisson is working right. it makes a poisson shape, it seems,
but it is _narrow_.
poisson for loss is particularly cheesy.
loss is given to ipfw as a real number 0..1.  ipfw translates it to 0..7fffffff.
i did not want to calculate P(x) for everything between 0 and 7fffffff, so i
do poisson for 0..7fff and shift it 16 bits before sending to kernel. i don't
think this makes a valid poisson dist. if you mess with the mean like that.

what next?
 try some restructuring. some things that made sense in vanilla dummynet
	became quite unwieldy as i kept adding things. making delay,loss,bw
	structs in the dn_pipe might clean things up (at the expense of having
	to say pipe->something->something).
 validity checks (eg mean 100 var 150)
 add more comments.
 test a lot. pretty much everything has had at least some pinging and
	"hmm, it seems to be doing what i want", but...