> not too convinced that poisson is correct (see corresponding dummynet > commit msg)
haven't touched yet

> the command line reading code has ballooned and could use cleaning in some
> way. since this is user code, i'm not _terribly_ concerned, but...

i played with it and made it more deterministic and i won't have to do tons
of "well, you specified this distribution but you didnt specify this arg"
now. it's not really as nice as i had imagined, but it is improved.

> should add more to 'ipfw pipe show'.

i added some. i tend to forget about it, 'cause it's not like i use this
much. i'm probably not going to bother with writing code to return the
entries in tables.

> more validity checking would be nice, but having the kernel catch it and
>  give a printf() and EINVAL is not bad.

added some. sometimes, i just let the kernel catch it.

--
what next? test a lot.