Initial import of the "capture" source into the testbed repository.
Note that this has a hardcoded value for NBPG, a define that's only
present in the digiboard headers.  In the interest of minimal source
code changes (heh, heh, heh) I left it alone. :p

Works on *bsd, and now Linux.  Manpage included at no charge,
though if someone wants to figure out why the manpage only
works under BSD ("Because Dave is an idiot" is not an acceptable
answer, true as it may be), that'd be cool.