Messing with frames. If you have your own defs.php3 file, make sure you pick up the changes in format from the default-defs.php3.
Anyway, I wasted over an hour trying to figure out why I couldn't get the
documents in right hand frame to go through the non-secure server. Turns
out that http://www.emulab.net and https://www.emulab.net are techically
different servers, and browsers have a "security" feature to prevent
"hijacked frames" whatever the hell that is. If the link to be placed in
the target frame is on a different server, it will open a new window
instead. So, by starting a http://www.emulab.net/welcome.html in the right
side, and then trying to put https://www.emulab.net/newproject.php3 in it,
a new window pops up.

Sigh. So, until I figure out a suitable work around, the Welcome frame
documentation is all going through the secure server (slowly of course).