New script: sshtb
sshtb is a _very_ simple shell script that runs ssh with a few commandline
parameters, which make it play nicer in an script environment. These
parameters can be changed with the '--with-ssh-args' argument, but default to:
'-q -o "BatchMode yes" -o "StrictHostKeyChecking no"'
All ssh calls now use this script.