First version of the testing software.
This is a functional check in.  To run the testing software do
something like:
	1. mkdir ~/testbed_test
	2. cd ~/testbed_test
	3. ~/testbed/testsuite/tbtest tbdb frontend
alternately:
	1. mkdir ~/testbed_test
	2. cd ~/testbed/testsuite
	3. tbtest -path ~/testbed_test tbdb frontend

A lot of stuff will be placed in testbed_test, including a configured
tree, an installed tree, a couple versions of the DB, and a whole
bunch of log files.

Interesting log files:

The main log file of interest is test.log which is a copy of the
output of tbtest.  The other log files are only useful if things fail.
Usually the fail message will include which log file to look in.  Log
files for tests are in tests/<test>/test.log .

What's missing:

1. Lot's of tests.  There's only one currently.

2. DB checking support.

3. A real README.  The current one is mainly a bunch of notes for the
real README.

4. Full support.  The current setup is only for testing the frontend.
Although there are some hooks for supporting full testing runs a lot
still needs to be done.