This commit contains two signifcant changes:
1. 'tb-set-hardware ... shark' and 'tb-set-hardware ... dnard' are now
functionally identical.  Previously only the former worked but both passed
the parser.

2. Assign will now exit very quickly in the case that, for a given virtual
nodes, there are no physical nodes that could match in type.  This should
never happen as the parser and assign_wrapper have checks that usually
prevent this.  However, in the case of problems in the code (such as #1)
this'll make it easier to debug.  In addition, as we add more types of
nodes and our estimates becoming increasingly inaccurate cases where this
might occur could slip in.  All calling code treats this identically to an
'insufficient resources' failure.