An attempt at making image creation an easy/automatic operation. HA! This uses the pxe booted freebsd kernel and MFS. In addition, I use the standard testbed mechanism of specifying a startup command to run, which will do the imagezip to NFS mounted /proj/<pid>/.... The controlling script on paper sets up the database, reboots the node, and then waits for the startstatus to change. Then it resets the DB and reboots the node so that it returns back to its normal OS. The format of operation is:
	create_image <node> <imageid> <filename>

Node must be under the user's control of course. The filename must
reside in the node's project (/proj/<pid>/whatever) since thats the
directory that is mounted by the testbed config software when the
machine boots. The imageid already exists in the DB, and is used to
determine what part of the disk to zip up (say, using the slice option
to the zipper). Since this operation is rather time consuming, it does
the usual trick of going to background and sending email status later.