Changes to sync protocol. The RPC27 spits out a bunch of silly status goo whenever it goes inactive for more than a couple of minutes. Also, require that we sync up so that we know for sure that the power control failed. Previously, I was blindly spitting out the command without being sure we had a command prompt.
Note, that we still have a problem with an active tip session messing
up power control since only one connection can be active at a time.
At present, there is no positive way to know that the connection was
refused since capture accepts the connection and then closes it if
there is an active tip session. I need to change the little handshake
so that there is a response Yes/Know response to the secret key
exchange.