Added Frisbee support to nfree. If the node is to be reloaded with Frisbee, puts it in the emulab-ops/reloading experiment and runs os_load for it, since the reload_daemon is not necessary with Frisbee.
In order for nfree to know what type of reload is desired, a new column was
added the the scheduled_reloads table, called reload_type. This gets set to
'frisbee' by sched_reload if called with the '-r' option.

nfree also now clears out the next_pxe_boot_path column for new nodes.