Interface change:
	Usage: os_load [-s | -w] [-r] [-i <imageid>] <node> [node ...]
        Usage: sched_reload [-f | -p] [-r] [-i <imageid>] <node> [node ...]

The imageid is now an optional argument. After continually forgetting
what imageid to use, or just plain forgetting the argument, and having
it try to load imageid pc53 on pcXX, I decided this interface was
bogus. With now imageid, select the default imageid for each node
provided. This is actually convenient since you can load multiple
types of nodes in one shot.