DB-ification major reformatting and commenting pass.
Now back to its old behavior, where frisbee reloads are handled no differently
than other reloads, because this difference is now taken care of in the
reload_daemon.

Note that the '-w' flag has been removed. This is due to an error that perl
reports on line 123, which, as far as I can tell, is not real. The code works
fine, anyway.