Add "routing" support with new TMCD command that looks at the routertype field in the nodes table. For eventual use when Mike is ready with gated stuff.
Modify the dohostnames stuff as per dicussion with Rob and
Mac. Instead semi-usless node-N names, set the name of the interface
to node-LinkName, where the linkname comes from the virt_lans table,
and is the name of the link/lan that interface is sitting on. This way
the user can actually figure out the interface when doing his routing
by assigning names to all links/lans in the NS file.

One major bogousity is that in order to maintain backwards
compatability, I've instituted a new hostnames commands (hostnamesV2).
We need a good way to deal with versioning!