Changes to how the backup files are named and saved. The existing naming scheme was not very intuitive and the names sucked. Also, I want to increase the frequency with which we run the backups, and I want to implement an automated roll so that we only keep about a months worth of history around.
Anyway, the new approach is to open up the index file and see what the
name of the current update file is. It has a numeric extension. Rename
the base log to base.XXX, and then snapshot the DB into backup.XXX So,
the combination of update.XXX and base.XXX is the DB history since the
last time the script was run. The file backup.XXX corresponds to the
DB at this point in time.

To restore (or track the changes of) a DB, simply take backup.XXX and
apply the changes that are stored in update.XXX+1 (which are the changes
made since backup.XXX was made). This should give you a DB that is the
same as backup.XXX+1. You can go back further, and just apply all the
subsequent update.XXX files.