Two new functions: getTrunks: Build a structure representing all of the trunks in the switch 	topolgy getTrunkPath: Find the shortest path between two switches, across the 	trunks found by getTrunks
Note that the getTrunkPath duplicates some work done in assign. However, we
don't currently have a good way to pass this information from assign to
snmpit, and we need to be able to figure this out anyway, in case snmpit
isn't being called as part of experiment swapin.

These functions should soon be used to pick which VLANs are allowed
across which trunks.