Code for picking which VLANs are allowed to go across which trunks. As much of it as possible switch-independant, and put in snmpit_lib. Should work for arbitrarily complicated switch toplogies, as long as there are not multiple trunks between two switches. (Multiple ports combined into one trunk are fine, however.)
As a result of this, VLAN creations and deletions now need to operate
on all switches, not just on the ones that have ports in the VLAN.
This is because the traffic may have to traverse switches that have no
ports in the VLAN to reach other switches that do.

Not called yet. I've done simple testing, but need to do more, as this
could get us into major trouble if it has bugs.