Actually call the code to pick which VLANs are allowed on which trunk.
When dealing with trunks, since they may be EtherChannels, we have to
make another lookup on the switch to get the ifIndex for the whole
channel - can't just use the ifIndex of one of the ports like you
can from the Cisco CLI.

These changes make snmpit slightly slower - it now has to get more information
from the switches when it's going to create or delete VLANs. However, this
is on the order of fractions of seconds, so it shouldn't be too noticable.