New script: schemacheck - Checks to see if the currently-running database matches the one in the checked-out source.
This now gets called as part of the 'boss-install' target, to guard
against installing software that is out-of-sync with the running
database. It is skipped if @prefix@ is not /usr/testbed, to avoid
getting in the way of development.

If you want to bypass this check, use the 'boss-install-force' target.
Use of this, however, is not recommended.