Rename exports_setup.proxy and console_setup.proxy to .in versions and remove the originals, so that we can run the files through configure.
NOTE: I wanted to keep the RCS history intact so I went over to the
CVS directory on moab and copied the ,v file to the new names, and
then did a normal cvs remove the originals. This keeps the RCS history
going without screwing up anyone. Not a recommended approach, but what
the hell.