Perl module for the event system.
This is accomplished with SWIG, so that we can basically just call the
C functions from perl, and avoid a parllel implementation. Some glue
code is required, particularly for callbacks, macros, and functions
that return data through arguments.

Currently lacks support for a few functions, but is functional enough
to re-write the example C programs.