Added two tables.
state_timeouts is for recording the maximum amount of time (in
seconds) that a node should be in a given state. (0 means no timeout)
The contents of the action column are not yet well-defined - in the
future, it may contain commands for dealing with stuck nodes, or
perhaps a keyword to indicate to the watchdog daemon what action
should be taken.

state_transitions contains a list of valid state transitions.