New script: stated
Watches for events sent by TMCD regarding the state of nodes. Records
this information in the database. Also watches for nodes that undergo
invalid state transitions, or stay in the same state for too long.
Right now, the only action it takes is to send email, but in the
future, will take action to 'unstick' nodes.

Not yet installed by default.